added mock server and routes

+++ config/setup.js
@@ -1,6 +1,13 @@
 import 'dotenv/config'
 import AuthHelper from '../helpers/auth.helper'
 import ConfigHelper from '../helpers/config.helper'
+import {start} from './server'
+
+const baseUrl = process.env.BASE_URL
+const port = process.env.PORT
+const isMock = baseUrl.includes('localhost') && baseUrl.includes(port)
+
+if (isMock) start(port)
 
 before(async function () {
   const authHelper = new AuthHelper()
@@ -9,6 +16,8 @@ before(async function () {
 })
 
 after(async function () {
-  const configHelper = new ConfigHelper()
-  await configHelper.delete()
+  if (!isMock) {
+    const configHelper = new ConfigHelper()
+    await configHelper.delete()
+  }
 })
